fix: move original index => sample, use index to redirect to pdf search
